// src/generated/arrow/flight/protocol/sql/FlightSql.ts
export interface CommandStatementQuery {
  /** The SQL syntax. */
  query: string;
  /** Include the query as part of this transaction (if unset, the query is auto-committed). */
  transactionId?: Buffer | undefined;
}
// ...
export interface CommandStatementUpdate {
  /** The SQL syntax. */
  query: string;
  /** Include the query as part of this transaction (if unset, the query is auto-committed). */
  transactionId?: Buffer | undefined;
}
// ...
export interface DoPutUpdateResult {
  /**
   * The number of records updated. A return value of -1 represents
   * an unknown updated record count.
   */
  recordCount: number;
}
// ...
export const CommandStatementQuery: MessageFns<CommandStatementQuery> = {
  encode(message: CommandStatementQuery, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.query !== "") {
      writer.uint32(10).string(message.query);
    }
    if (message.transactionId !== undefined) {
      writer.uint32(18).bytes(message.transactionId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CommandStatementQuery {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCommandStatementQuery();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.query = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.transactionId = Buffer.from(reader.bytes());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CommandStatementQuery {
    return {
      query: isSet(object.query) ? globalThis.String(object.query) : "",
      transactionId: isSet(object.transactionId)
        ? Buffer.from(bytesFromBase64(object.transactionId))
        : isSet(object.transaction_id)
        ? Buffer.from(bytesFromBase64(object.transaction_id))
        : undefined,
    };
  },

  toJSON(message: CommandStatementQuery): unknown {
    const obj: any = {};
    if (message.query !== "") {
      obj.query = message.query;
    }
    if (message.transactionId !== undefined) {
      obj.transactionId = base64FromBytes(message.transactionId);
    }
    return obj;
  },

  create(base?: DeepPartial<CommandStatementQuery>): CommandStatementQuery {
    return CommandStatementQuery.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<CommandStatementQuery>): CommandStatementQuery {
    const message = createBaseCommandStatementQuery();
    message.query = object.query ?? "";
    message.transactionId = object.transactionId ?? undefined;
    return message;
  },
};
// ...
export const CommandStatementUpdate: MessageFns<CommandStatementUpdate> = {
  encode(message: CommandStatementUpdate, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.query !== "") {
      writer.uint32(10).string(message.query);
    }
    if (message.transactionId !== undefined) {
      writer.uint32(18).bytes(message.transactionId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CommandStatementUpdate {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCommandStatementUpdate();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.query = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.transactionId = Buffer.from(reader.bytes());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CommandStatementUpdate {
    return {
      query: isSet(object.query) ? globalThis.String(object.query) : "",
      transactionId: isSet(object.transactionId)
        ? Buffer.from(bytesFromBase64(object.transactionId))
        : isSet(object.transaction_id)
        ? Buffer.from(bytesFromBase64(object.transaction_id))
        : undefined,
    };
  },

  toJSON(message: CommandStatementUpdate): unknown {
    const obj: any = {};
    if (message.query !== "") {
      obj.query = message.query;
    }
    if (message.transactionId !== undefined) {
      obj.transactionId = base64FromBytes(message.transactionId);
    }
    return obj;
  },

  create(base?: DeepPartial<CommandStatementUpdate>): CommandStatementUpdate {
    return CommandStatementUpdate.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<CommandStatementUpdate>): CommandStatementUpdate {
    const message = createBaseCommandStatementUpdate();
    message.query = object.query ?? "";
    message.transactionId = object.transactionId ?? undefined;
    return message;
  },
};
// ...
export const DoPutUpdateResult: MessageFns<DoPutUpdateResult> = {
  encode(message: DoPutUpdateResult, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.recordCount !== 0) {
      writer.uint32(8).int64(message.recordCount);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): DoPutUpdateResult {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseDoPutUpdateResult();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.recordCount = longToNumber(reader.int64());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): DoPutUpdateResult {
    return {
      recordCount: isSet(object.recordCount)
        ? globalThis.Number(object.recordCount)
        : isSet(object.record_count)
        ? globalThis.Number(object.record_count)
        : 0,
    };
  },

  toJSON(message: DoPutUpdateResult): unknown {
    const obj: any = {};
    if (message.recordCount !== 0) {
      obj.recordCount = Math.round(message.recordCount);
    }
    return obj;
  },

  create(base?: DeepPartial<DoPutUpdateResult>): DoPutUpdateResult {
    return DoPutUpdateResult.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<DoPutUpdateResult>): DoPutUpdateResult {
    const message = createBaseDoPutUpdateResult();
    message.recordCount = object.recordCount ?? 0;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}